+++ bin/utils/coverage.sh
@@ -1,11 +1,31 @@
 #!/usr/bin/env bash
 set -e
 
-err_report() {
-    echo "ERROR: on line $*: $(cat $0 | sed $1!d)" >&2
+
+function get_stack () {
+   STACK=""
+   local i message="${1:-""}"
+   local stack_size=${#FUNCNAME[@]}
+   # to avoid noise we start with 1 to skip the get_stack function
+   for (( i=1; i<stack_size; i++ )); do
+      local func="${FUNCNAME[$i]}"
+      [[ $func = "" ]] && func=MAIN
+      local linen="${BASH_LINENO[$(( i - 1 ))]}"
+      local src="${BASH_SOURCE[$i]}"
+      [[ "$src" = "" ]] && src=non_file_source
+
+      STACK+=$'\n'"   at: $func $src "$linen
+   done
+   STACK="${message}${STACK}"
 }
 
-trap 'err_report $LINENO' ERR
+err_report() {
+    get_stack
+    echo "ERROR: on line $*: $(cat $0 | sed $1!d)" >&2
+    exit 8
+}
+
+trap 'get_stack' ERR
 
 uid=$(id -u)
 
@@ -14,13 +34,11 @@ if [ "${uid}" -eq 0 ]; then
     exit
 fi
 
-pwd=$(pwd)
-
 install="$1"
 
 opts=( "${@:2}" )
 
-. "${pwd}/bin/scripts/load.sh"
+. "$(pwd)/bin/scripts/load.sh"
 
 ## Builds binaries and creates coverage info
 ## Usage: ${script} <option> [optional]
@@ -42,6 +60,10 @@ case "${install}" in
         opts=( "${@:2}" )
         ;;
 
+    "--help") # [optional] shows command desctiption
+        help
+        ;;
+
     *)
         help
         ;;
@@ -58,7 +80,7 @@ for opt in ${opts[@]}; do
             dir=${opt#*=}
             ;;
 
-        "--clean") # [optional] cleans up directories before build
+        "--clean") # [optional] cleans up directories before coverage
             clean="--clean"
             ;;
 
@@ -78,15 +100,15 @@ for opt in ${opts[@]}; do
             silent="--silent"
             ;;
 
-        "--valgrind") # [optional] runs using valgrind (disables --sanitize on build)
+        "--valgrind") # [optional] runs using valgrind
             valgrind="--valgrind"
             ;;
 
-        "--callgrind") # [optional] runs using valgrind with tool callgrind (disables --sanitize on build)
+        "--callgrind") # [optional] runs using valgrind with tool callgrind
             callgrind="--callgrind"
             ;;
 
-        "--debug") # [optional] runs using debug memory debug info
+        "--debug") # [optional] runs using debug messaging
             debug="--debug"
             ;;
 
@@ -101,7 +123,7 @@ for opt in ${opts[@]}; do
     esac
 done
 
-if [ "${silent}" == "--silent" ]; then
+if [[ "${silent}" == "--silent" ]]; then
     exec 2>&1 >/dev/null
 fi
 
@@ -111,21 +133,10 @@ if [[ ! "${dir}" == "" ]]; then
     build="${dir}"
 fi
 
-[ ! -d "${build}" ] && mkdir "${build}"
-
-if [ "${clean}" == "--clean" ]; then
-    rm -rf "${build}"
-fi
-
-find "${pwd}" -type f -name "callgrind.out.*" -delete
-find "${pwd}/src" -type f -name "*.s" -delete
-find "${pwd}/tests" -type f -name "*.s" -delete
-
-cmake=$(get-cmake)
-
-if [[ "${cmake}" == "" ]]; then
-    echo cmake not found. please run "$(pwd)/bin/utils/install.sh" --cmake
-    exit 8
+if [[ "${clean}" == "--clean" ]]; then
+    if [[ -d "${dir}" ]]; then
+        rm -rf "${dir}"
+    fi
 fi
 
 targets=( $(get-source-targets ${source}) )
@@ -134,34 +145,47 @@ if [[ "${targets[@]}" == "" ]]; then
     if [[ "${help}" == "--help" ]]; then
         help
     fi
-    echo ERROR
+    echo no source targets: ${source} >&2
     exit 8
 fi
 
-[ ! -d "${build}" ] && mkdir "${build}"
+cmake=$(get-cmake)
 
-output="${pwd}/output"
-[ ! -d "${output}" ] && mkdir "${output}"
+if [[ "${cmake}" == "" ]]; then
+    echo cmake not found. please run "$(pwd)/bin/utils/install.sh" --cmake
+    exit 8
+fi
+
+output="$(pwd)/output"
+[[ ! -d "${output}" ]] && mkdir "${output}"
 
 for target in ${targets[@]}; do
+    if [[ -f "${build}/${target}" ]]; then
+        rm "${build}/${target}"
+    fi
     if [[ -f "${output}/log-${target}.txt" ]]; then
         rm "${output}/log-${target}.txt"
     fi
+
+    if [[ -f "${build}/${target}.info" ]]; then
+        rm "${build}/${target}.info"
+    fi
 done
 
-if [[ -f "${build}/${target}.info" ]]; then
-    rm "${build}/${target}.info"
-fi
-
 coverage=( "*.gcda" "*.gcno" "*.s" "*.i" "*.o" "*.info" )
+
 for f in ${coverage[@]}; do
-    find "${build}" -type f -name "${f}" -delete
+    if [[ -d "${build}" ]]; then
+        find "${build}" -type f -name "callgrind.out.*" -delete
+        find "${build}" -type f -name "*.s" -delete
+        find "${build}" -type f -name "${f}" -delete
+    fi
 done
 
 export LCOV_PATH=$(which lcov)
 export GENHTML_PATH==$(which genhtml)
 export MAKEFLAGS=-j8
-export LD_LIBRARY_PATH="${pwd}/lib"
+export LD_LIBRARY_PATH="$(pwd)/lib"
 
 ${cmake} \
     -DCMAKE_EXPORT_COMPILE_COMMANDS:BOOL=TRUE \
@@ -172,14 +196,14 @@ ${cmake} \
     -DGENHTML_PATH=${GENHTML_PATH} \
     -DCODE_COVERAGE:BOOL=TRUE \
     $(cmake-options) \
-    -S"${pwd}" \
+    -S"$(pwd)" \
     -B"${build}" \
     -G "Ninja" 2>&1 >/dev/null
 
 for target in ${targets[@]}; do
     echo building ${target}
     echo options $(cmake-options)
-    if [ "${silent}" == "--silent" ]; then
+    if [[ "${silent}" == "--silent" ]]; then
         ${cmake} --build "${build}" --target "${target}" 2>&1 >/dev/null || (echo ERROR: "${target}" && exit 1)
     else
         ${cmake} --build "${build}" --target "${target}" || (echo ERROR: "${target}" && exit 1)
@@ -188,25 +212,28 @@ for target in ${targets[@]}; do
         c-*) ;& main-*) ;& test-*)
             timeout --foreground 180 $(cmake-valgrind-options) "${build}/${target}" 2>&1 >"${output}/log-${target}.txt" || (echo ERROR: "${target}" && exit 1)
             lcov --capture --directory "${build}/" --output-file "${build}/${target}.info" &>/dev/null
-            lcov --remove "${build}/${target}.info" "${pwd}/.deps/*" -o "${build}/${target}.info"
+            lcov --remove "${build}/${target}.info" "$(pwd)/.deps/*" -o "${build}/${target}.info"
             ;;
         *)
             ;;
     esac
 done
 
-find "${build}" -type f -name "*.info" -exec echo -a {} \; | xargs lcov -o "${build}/lcov.info"
+files=$(find "${build}" -type f -name "*.info" -exec echo {} \;)
+if [[ ! "${files[@]}" == "" ]]; then
+    find "${build}" -type f -name "*.info" -exec echo -a {} \; | xargs lcov -o "${build}/lcov.info"
+fi
 
 main=$(find "${build}" -type f -name "*.s" -exec echo {} \;)
 for i in ${main[@]}; do
-    path="${pwd}/$(echo $i | sed -n -e 's/^.*.dir\/\(.*\)$/\1/p')"
+    path="$(pwd)/$(echo $i | sed -n -e 's/^.*.dir\/\(.*\)$/\1/p')"
     cp "${i}" "${path}"
 done
 
-if [ "${silent}" == "--silent" ]; then
+if [[ "${silent}" == "--silent" ]]; then
     exec 1>&2 2>&-
 fi
 
 [[ $SHLVL -gt 2 ]] || echo OK
 
-cd "${pwd}"
+cd "$(pwd)"
